move list bonders related stuff to packages
